fix(transport): port gemini rate limit retry and API key fallback to AcpTransport
PR #516 migrated gemini-cli from LegacyTransport to AcpTransport but
did not port the rate limit retry logic, causing immediate failure on
OAuth quota exhaustion without any model switch or API key fallback.

Changes:
- Extract gemini retry utility functions from LegacyTransport into
  transport_gemini_retry.rs (shared between Legacy and ACP transports)
- Add retry loop to AcpTransport::execute for gemini-cli:
  attempt 1 → original model, attempt 2 → gemini-3.1-pro-preview,
  attempt 3 → gemini-3-flash-preview, then API key fallback
- Extract execute_acp_attempt helper to enable retry without code
  duplication in the spawn_blocking closure
- Move inline transport tests to tail file to keep transport.rs under
  800-line monolith limit

Closes #519
